Inital commit of dataFascia version 2 code base.
This commit is to create a template and general directory structure for
version 2 of the Sapphire code for dataFascia. Some key elements of this
check-in:

 * The focus is to start from the REST API server and bind back to the ETL
   code
 * Make use of Dropwizard code for the REST server
 * Make use of Lombok to avoid writing boiler plate code
 * Integrate Atmosphere to provide Websockets for event streaming
 * Make use of TestNg for testing

This check-in has two dummy resources to capture our current thinking on
how to develop the API server. We want to separate our internal data model
from the resource model. This will allow us to insert computational code
(specifically security related) between the internal and external view of
data.