Add template for Emerge applications in source code
All Emerge applications will be consolidated under one directory
emerge-apps under which eadh individual Emerge application will reside.
The first application is the conversion of df data to Emerge CSV format
which is under emerge-csv. All models needed for the Emerge applications
will be defined in emerge-models.

The model in this commit is the Demographic model which maps to the
Demographic data CSV needed by Emerge.