Add string-template to format name
Based on requirements, the rules for generating a name from individual
components will be different. Rather than encoding these rules all over
the place, it would be better to use the concept of a formatter to change
the order necessary on demand. StringTemplate is a pretty good library for
this purpose.