Move emerge-csv to subcommand of shell command
After this change, the project creates the artifact
df-shell-${project.version}-bin.tar.gz containing a shell script that
runs a Java application with its dependencies.

To run the a specific process, execute the shell command:

  dfp <command-name> [options]

For example, to generate the demographic-csv:

  dfp demographic-csv -apiURI ... -csvFile ... ...

This package will club all shell based commands.