Create a websocket API resource that delivers streaming Kafka messages.
Based on a requirement that API clients mus be able to subscribe to
streaming data.

This is a resource hosted inside of Dropwizard, using Atmosphere library
for handling the websocket session. The resource is configured by
central Dropwizard properties, creates a Kafka consumer which upon a
client request, subscribes to a requested Kafka topic and provides
those messages over websocket to the client.

Path has been changed and reimplemented under ManagedService in SocketResource
to match KafkaResource and fix a bug where SocketResource would pick up inbound
connections intended for KafkaResource.

Shiro config has been changed to disable authentication for /websocket/**

This commit addresses fundamental operation of the resource and does not
address granular use cases. For instance, only one configuration property for
Kafka Consumers is exposed as a GET parameter, when more will likely be needed
to handle edge cases like resuming broken connections.

This commit does not include testing for this functionality.

This commit addresses Jira ticket SAP-397