Add ability to return JSON schema for data models
This is useful for developers to understand the structure of our data
models. Most of the schema generation features comes from an external
package (jackson-module-jsonSchema) which utilises the Jackson annotations
in the models to render the information.

Three ways to fetch the schema:

 * /schema : Will fetch JSON schema for all objects
 * /schema/<ModelName> : Will fetch the JSON schema for the model
 * /<ModelName>/schema : Will fetch the JSON schema for the model

Please note the two ways the schema for a single model can be fetched.
This will start to play into when full blown HATEOS is implemented.

There are two interesting side-effects which are part of this commit:

 * Default package gets hard coded into VersionResource. Reason is that
   each resource is now being constructed via Guice and a null argument
   constructor is needed. The default package being picked from
   configuration was a place holder so no impact on functionality
 * All java files in com.datafascia.api.resources will automatically be
   loaded in APIService. This way when new resources are added, one does
   not have to make a change to this file to register the resource anymore