Add feature to manage mapping between URN ID namespace and classes
The key idea behind this feature is to be able to annotate model classes
with their ID namespaces. For for example if the Patient model identifier
URN is in the namespace "df-patient-id", the annotation
@IDNamespace("df-patient-id") can be added to the Patient object. The
URNMap can then be passed all package names where it searches for classes
with the annotation and initializes the two way lookup map. The URNMapTest
shows how to initialize the map.

Class creation with {{...}} initialization seems to cause the reflections
package some problems in considering the classes annotated because of
sub-classing, but not to find the annotations themselves. This seems to
happen for test code only. Thus the additional check for null annotation.

Log the classes found at INFO level so the log file can be checked for
consistency.

NOTE: IDNamespace is a 'type' level annotation