Refactors Accumulo authorization handling
Hard coding "System" in each place is probably not a very good idea. This refactors the
the Accumulo authorization as follows:

  * The 'User' credential will have the authorization allowed
  * This authorization is then passed via the resources to the dao functions

NOTE: The authorization is hard coded to "System" in User, but it is one place and once real
security is integrated in, that is a good place to pass the parameter anyway.