Completes basic tests for remaining core models
This largely contains implementation of very basic tests for the models in
core that are not yet covered. The most basic test uses the one test method
in ModelTestBase, and at least this test has been implemented for all models.

Null checks and long string tests were also added to NameTest.