Add template method to save entries to a row
This method executes the steps to write to Accumulo including creating a
batch writer, writing a mutation, and catching exceptions. These steps
would otherwise be boilerplate code if implemented outside this method.
The template method creates a mutation and passes it to a caller
provided function that must put write operations into the mutation.