Pages patients back for API call
Instead of sending back the complete list of patients, the client
and server can now control the upper limit that can be returned. It
makes use of the PagedCollection model to control the total number
sent back.

NOTE: the api-client has also changed to accept the new parameters

Also changed the patient ids in test to not have the URN. Need to
rethink our in-memory strategy a bit.