Creates scenario tests
The scenario tests are used to verify the operation of the system.

Scenarios are defined in a scenario file in src/test/resource/scenarios.
Each scenario file contains a number of steps and each step has 3 components
separated by the vertical bar, "|".  The components are:

  1. The HL7 message file to load
  2. The ID of the patient in the HL7 message
  3. The time interval to wait in seconds for message to propagate through the
     system before checking the results

The HL7 file is introduced to the system via the MLLP server. The scenario assumes
the services are all available and running and the paremeters are passed via
environment variables

fixes bug: SAP-345