Remove duplicate HL7 transform code
HL7 version 2 is a backwards compatible standard, for the most part. It
is possible to use a HAPI message structure to parse a message of the
same type from an earlier version of the standard. Force HAPI to create
message structures for HL7 2.4. The model classes will be compatible
with messages from previous versions.

See SAPP-95