Adds patients to test instance.
The PatientIT was originally written with Opal patients being tested. Since then we've moved to the Patients table with the new schema. The test silently broke as it was relying on a switch statement to find patients, so if 0 patients, the test was effectively skipped. Now I'm manually adding the expected patients and then checking to be sure all 6 were found.