Changes scenario description from text to JSON
The scenario and step description file will only get more complicated over time.
As the code for extracting description shows that having a cleaner definition
extendable over time will keep the code simple. Since the data structure syntax
has been JSON so far, it makes sense to to change the scenario description
language also to the same.

The transfer of information is almost one to one except for the fact that the patient
and institution identifier has been moved to the scenario level rather than step level.

Also includes changing names of scenarios to be more user friendly.

Implements SAPP-83