Renames name member in Observation model to code.
The Observation model had two members that represented the same thing.
FHIR's Observation model has a member called code that pulls in OBX 3.1.
This member was present in our observation model, along with a field called
"name" which was being used instead of code to hold OBX 3.1, but is not
present in the FHIR DSTUv2 model.

All methods using name were changed to use code and the name member has
been erased.

Addresses JIRA ticket SAPP-132