Adds Observation create and integration test for API Server.
API server doesn't like multi-parameter create calls, so a single parameter
save method was added to the Observation repository that works by pulling
the Encounter reference from the Observation that is handed from the client.
This means that the client must know the internal ID of the referred
encounter and include that reference in the Observation before passing it
to the API's create method.

Addresses ticket SAPP-134