Upgrade to commons-codec 1.10
Before this change, the shell module packaged commons-codec 1.4, in
which the Base64.encodeBase64String method appended CR LF to its return
value. The HAPI FHIR client sent HTTP requests that were malformed
because this inserted an empty line in middle of the HTTP headers. The
HAPI FHIR client depends on a newer commons-codec version in which the
behavior of this method is changed to not append CR LF.