Adds 2-argument search to resources for MedicationAdministration and MedicationPrescription.
This adds the capability to retrieve a sepcific MedicationAdministration or MedicationPrescription
resource by our API. Both of these resources, when retrieving by ID, require also the ID of the
parent Encounter resource, based on how our data is structured internally. Due to this,
the "read" operation in the API resource is implemented here as a two argument search that
returns a list with 0 or 1 members, as a two argument read is not supported in HAPI-FHIR.

Addresses ticket SAPP-134.