Changes Rules MedsSet result to a list.
Duplicate firings were already occuring with erroneous rules
for which we're still waiting for clarification to fix.

Further duplicate firings from events will come once events are
able to hold lists of RxCuiIN (ingredients).

Therefore the result of the rule firing will be now stored in a list of
MedsGroup elements, instead of an individual element.

Addresses SAPP-137