Return HTTP status 404 if resource not found. Enforce consistent search behavior.
Reads on resources must throw a resource not found exception if not found.
Some resources we want to retrieve by fields that are not the resource ID,
although we expect 1 or 0 results these are still searches, and searches
must always return a list, empty if no match is found. This behavior
is also enforced in this commit.

Addresses ticket SAPP-149.