Fixes meds set read in Anticoagulation for Med Orders.
Meds set names are stored in one of the identifiers in MedicationOrder.identifier.
This changes anticoagulated to use this method of getting the meds set from
MedicationOrder.

This addresses SAPP-263 and fixes a previous NPE on line 73.

The second half of Anticoagulated, the question of retriving med names from
Administrations in this case is not addressed in this commit.