Discard Procedures, Locations that lack identifiers
Procedure repository was throwing NPEs because it can't save a Procedure
without an Identifier field from which to pull.

This is the problem in SAPP-311 and resembles the problem that used
to be in SAPP-295 for Practitioners.

This commit fixes SAPP-311 by a null check for Procedures and then
includes a prophylactic fixes for another resource that could
possibly cause the same problem. In both cases this change logs
resources where an ID cannot be created and discards the unsavable
resource.