Fixes MedicationAdministration MedsSet access.
Meds Sets are pulled from identifier fields in MedicationAdministrations.
A given MedicationAdministration can have many identifiers, the first
being a non-meds-set ID and then multiple meds sets.

Lots of code attempted to test for meds set by reading the zeroth
med admin identifier. This code will fail when the desired identifier
is present but not in the expected place.

This code corrects this improper method of access where found and replaces
it with a standard method for searching meds sets.