Adds precision specifier to BigDecimal division.
Apparently BigDecimal.divide will throw an ArithmeticException if
a division results in a non-terminating decimal. In order to avoid
these results, precision must be specified with a 3 argument divide
method. The second argument is the "scale" which is the max digits
of precision after the decimal point.

Replaced all bigdecimal division with the 3-arg call and set precision
to 10 places in case the call is ever pasted where increased precision
is important.