For Pain Goal from CPOT, remove acceptable level of pain assessment requirement.
Customer reported a patient with pain goal 11, cpot observation of 4, no other
numeric or verbal pain observations, no acceptable level of pain assessment observations.

The code as written was requiring an acceptable level of pain assessment observation with
a certain value or returning 11. However, this disagrees with the pseudocode which states
that this requirement for this specific observation type only applies to verbal and numerical
pain types, not cpot.

This commit makes the requirement of an acceptable level of pain assessment specific to
numerical and verbal and excludes it from CPOT processing.

Addresses SAPP-445.