Temporarily return Propofol and Dexmedetomidine to PD sedatives.
Although they don't belong there, old harmevidence instances sit
on the database that are crashing the api server when they try to
serialize.

We need to add the drug names back to the schema so that they can
serialize, so harmevidence can run and eventually the two incorrect
drugs will be removed on update.

Once that happens, this commit can be rolled back.