Temporarily return PC and AC to tidal volume json schema enum.
Although they don't belong there, old harmevidence instances sit
on the database that are crashing the api server when they try to
serialize.

We need to add the old tidal volume names back to the schema so
that they can serialize, so harmevidence can run and eventually the
two incorrect volumes will be removed on update.

Once that happens, this commit can be rolled back.