Handle null dates in med admins.
When the web services have a null date, they pass the string \/Date(-62135568000000-0800)\/ and we read that in literally. David is modifying the service to ommit the time if there is no time associated with it, so we'll no longer get those strings. Currently, Due administrations (which we don't record) are the only place we see this, but I still want to handle the case that they give us an admin without an effective time.