Improves efficiency of the med admin processor.
A lot of simple things, like instead of constantly refetching the encounter, fetch it once and pass it along. Also, since no orders or admins can be saved if the encounter isn't in FHIR, a lot is shortcut by checking the encounter's existence from the getgo and moving on. Same deal with orders, fetching them once and passing them in to the admins instead of having each admin processing step refetch the order.