Process HL7 from messages saved by encounter
After this change, the ProcessHL7 NiFi processor expects each input
FlowFile to have an encounterIdentifier attribute. ProcessHL7 uses this
attribute to retrieve and process HL7 messages for that encounter.
ProcessHL7 remembers the last processed message ID, so each message is
processed only once, even when the patient is transferred out of and
into the ICU.