Removes active or draft order check for intermittent anticoagulant administrations.
An active or draft order check is necessary for the new isActivelyInfusing tests
we have been given to determine if continuous infusions are stopped.

In the case of AnticoagulationImpl this check was inappropriately used for both
intermittent and continuous administrations, when it should only apply for
continuous administrations.

This check was removed and tests were modified to make sure that intermittent
anticoagulant administrations are not falsely disregarded. The check is still
called for continuous infusions in the isActivelyInfusing call.

Addresses SAPP-558