Upgrade to HAPI-FHIR 1.6-SNAPSHOT
Corresponding with James Agnew over profiling discovering slowness in
the wereBack method, a change was pushed to HAPI-FHIR that removes
O(n^2) logic in the bundle parsing code.

This O(n^2) loop was consuming the majority of CPU cycles in production,
making the use of a snapshot version of HAPI-FHIR the probable
"lesser of two evils".