Moves date parsing code to epicDateToInstant.
A change to date parsing was erroneously left in an unused method.
The method has been removed and this moves the code to epicDateToInstant
and adds a unit test for the method.

This unit test can be easily repurposed for parsing hard-to-read
date formats to plaintext when troubleshooting raw data.

Addresses SAPP-576